Completed get HW info for Linux
git-svn-id: svn+ssh://192.168.1.105/Users/mikael/icl-repository/trunk@813 763894fe-7313-0410-9b4e-9763914fbcec

+++ include/ic_hw_info.h
@@ -35,14 +35,14 @@ void ic_get_cpu_info(guint32 *num_processors,
 struct ic_mem_info
 {
   guint32 numa_node_id;
-  guint64 memory_size;
+  guint32 memory_size;
 };
 typedef struct ic_mem_info IC_MEM_INFO;
 
 void ic_get_mem_info(guint32 *num_numa_nodes,
-                     guint64 *total_memory_size,
+                     guint32 *total_memory_size,
                      IC_MEM_INFO **mem_info);
 
 void ic_get_disk_info(gchar *directory_name,
-                      guint64 *disk_space);
+                      guint32 *disk_space);
 #endif
